Add cron_d resource from the cron cookbook
This was the #1 requested resource that we already had written. Everyone uses it and we should ship it. It's not a replacement for the existing cron cookbook, but just a different way of managing cron. If people want to manage the whole file or individual snippets we'll give them the option, just as we've always done by having the core resource and this resource.

Signed-off-by: Tim Smith <tsmith@chef.io>